refactor repl.go
